feat: 설문 응답하기 기능 개발

+++ src/main/java/com/bluewhaletech/Ourry/dto/ChoiceDTO.java
@@ -11,5 +11,5 @@ public class ChoiceDTO {
     private String detail;
 
     @NotBlank
-    private Long seq;
+    private int seq;
 }
